+++ libs/ferriskey-domain/Cargo.toml
@@ -11,6 +11,7 @@ serde = { version = "1.0.228", features = ["derive"] }
 utoipa = { version = "5.4.0", features = ["chrono", "uuid"] }
 uuid = { version = "1.20.0", features = ["serde", "v4", "v7"] }
 thiserror = "2.0.12"
+tracing = "0.1"
 
 [dev-dependencies]
 serde_json = "1.0.149"
