+++ libs/ferriskey-domain/src/common/mod.rs
@@ -1,1 +1,2 @@
 pub mod app_errors;
+pub mod policies;
