+++ libs/ferriskey-domain/src/role/mod.rs
@@ -1,4 +1,5 @@
 pub mod commands;
 pub mod entities;
+pub mod permission;
 pub mod ports;
 pub mod value_objects;
